Finally refactoring into more modular style.
