+++ src/clr/mod.rs
@@ -1,0 +1,11 @@
+//mod.rs
+mod metahost;
+mod runtime_info;
+mod runtime_host;
+mod c_api;
+mod assembly;
+
+pub use self::assembly::Assembly;
+pub use self::metahost::CLRMetaHost;
+pub use self::runtime_info::CLRRuntimeInfo;
+pub use self::runtime_host::CLRRuntimeHost;
