+++ src/lib.rs
@@ -1,223 +1,10 @@
 extern crate libc;
 extern crate winapi;
-use winapi::ctypes::{c_void, c_char, wchar_t };
-use winapi::shared::winerror::{HRESULT, E_POINTER};
-use winapi::shared::ntdef::HANDLE;
-use std::ptr;
-use std::ffi::CString;
-
 extern crate widestring;
-use widestring::{WideString};
 
-#[repr(C)]
-struct CAPIResult {
-	hr: HRESULT, 
-	ok: bool, 
-	c_ptr: *mut c_void, 
-	ws_ptr: *mut u16
-}
+mod clr;
+use clr::CLRMetaHost;
 
-extern "C" {
-	//CLRMetaHost functions
-	fn CLRMetaHost_new()                                                  -> CAPIResult;
-	fn CLRMetaHost_get_runtime(host: *mut c_void, version: *const c_char) -> CAPIResult; 
-	
-	//CLRRuntimeInfo functions
-	fn CLRRuntimeInfo_is_loadable(runtime_info: *mut c_void) -> bool;
-	fn CLRRuntimeInfo_is_loaded(runtime_info: *mut c_void) -> bool;
-	fn CLRRuntimeInfo_is_loaded_from_handle(runtime_info: *mut c_void, process_handle: *mut c_void) -> bool;
-	fn CLRRuntimeInfo_is_started(runtime_info: *mut c_void) -> bool;
-	fn CLRRuntimeInfo_get_clr_runtime(runtime_info: *mut c_void) -> CAPIResult;
-	fn CLRRuntimeInfo_load_error_string(runtime_info: *mut c_void, hr: HRESULT) -> CAPIResult;
-	fn CLRRuntimeInfo_load_library(runtime_info: *mut c_void, dll_name: *const c_char) -> CAPIResult;
-	
-	//utility method
-	fn CAPI_free_error_string(error_string: *mut c_char);
-	
-	//CLRRuntimeHost functions
-	fn CLRRuntimeHost_stop(clr_host: *mut c_void) -> CAPIResult;
-	fn CLRRuntimeHost_start(clr_host: *mut c_void) -> bool;
-	fn CLRRuntimeHost_load_assembly(clr_host: *mut c_void, clr_info: *mut c_void, assembly_name: *const c_char) -> CAPIResult;
-	
-	//Assembly functions
-	fn Assembly_release(assembly: *mut c_void) -> CAPIResult;
-	fn Assembly_get_type(assembly: *mut c_void) -> CAPIResult;
-	
-	//Type functions
-	fn Type_call_static_method(typ: *mut c_void, method_name: *const c_char) -> CAPIResult;
-	fn Type_create_instance(typ: *mut c_void) -> CAPIResult;
-	
-	//Object instance functions
-	fn Object_invoke(obj: *mut c_void, method_name: *const c_char) -> CAPIResult;
-	fn Object_free(obj: *mut c_void) -> CAPIResult;
-}
-
-#[repr(C)]
-pub struct CLRMetaHost {
-	internal_ptr: *mut c_void 
-}
-
-impl CLRMetaHost {
-	pub fn new() -> CLRMetaHost {
-		let res = unsafe {CLRMetaHost_new()};
-		
-		let ptr = match res.ok {
-			true => res.c_ptr, 
-			false => { panic!("CLRMetaHost::new call failed with HRESULT: {:?}", res.hr); }
-		};
-		assert!(!ptr.is_null());
-		CLRMetaHost{ internal_ptr: ptr }
-		
-	}
-	
-	pub fn get_runtime_info(&self, version: &str) -> CLRRuntimeInfo {
-		assert!(!self.is_null());
-		let cs_version = CString::new(version).unwrap();
-		let res = unsafe { CLRMetaHost_get_runtime(self.internal_ptr, cs_version.as_ptr()) };
-		let ptr = match res.ok {
-			true => res.c_ptr, 
-			false => {panic!("get_runtime_info call failed with HRESULT: {:?}", res.hr);}
-		};
-		assert!(!ptr.is_null());
-		CLRRuntimeInfo::new(ptr)
-	}
-	
-	pub fn is_null(&self) -> bool {
-		self.internal_ptr.is_null()
-	}
-}
-
-#[repr(C)]
-pub struct Assembly(*mut c_void);
-
-impl Assembly {
-	fn new(c_ptr: *mut c_void) -> Assembly {
-		Assembly(c_ptr)
-	}
-}
-
-impl Drop for Assembly {
-	fn drop(&mut self) {
-		unsafe {
-			Assembly_release(self.0);
-		}
-	}
-}
-
-#[repr(C)]
-pub struct CLRRuntimeInfo {
-	internal_ptr: *mut c_void,
-	loaded: bool, 
-	loadable: bool, 
-	started: bool
-}
-
-impl CLRRuntimeInfo {
-	pub fn new(ptr: *mut c_void) -> CLRRuntimeInfo {
-		CLRRuntimeInfo {
-			internal_ptr: ptr, 
-			loaded: false, 
-			loadable: false, 
-			started: false
-		}
-	}
-	pub fn is_loadable(&mut self) -> bool {
-		if !self.loadable {
-			self.loadable = unsafe { CLRRuntimeInfo_is_loadable(self.internal_ptr)};
-		}
-		self.loadable
-	}
-	
-	pub fn is_loaded(&mut self) -> bool {
-		if !self.loaded {
-			self.loaded = unsafe { CLRRuntimeInfo_is_loaded(self.internal_ptr) };
-		}
-		self.loaded
-	}
-	
-	pub fn is_started(&mut self) -> bool {
-		if !self.started {
-			self.started = unsafe {CLRRuntimeInfo_is_started(self.internal_ptr)};
-		}
-		self.started
-	}
-	
-	pub fn is_loaded_from_handle(&mut self, process_handle: HANDLE) -> bool {
-		unsafe {
-			CLRRuntimeInfo_is_loaded_from_handle(self.internal_ptr, process_handle)
-		}
-	}
-	
-	pub fn get_clr_host(&self) -> Result<CLRRuntimeHost, HRESULT> {
-		let res = unsafe {
-			CLRRuntimeInfo_get_clr_runtime(self.internal_ptr)
-		};
-		match res.ok {
-			true => Ok(CLRRuntimeHost{internal_ptr: res.c_ptr, started: false, assemblies: Vec::new()}), 
-			false => Err(res.hr)
-		}
-	}	
-	
-	pub fn get_error_string(&self, hr: HRESULT) -> Result<WideString, String> {
-		let result = unsafe {CLRRuntimeInfo_load_error_string(self.internal_ptr, hr) };
-		match result.ok {
-			true => {
-				Ok( unsafe {WideString::from_ptr(result.ws_ptr, result.hr as usize) })
-			}, 
-			false => Err("Couldn't load the resource string. Check that you have the correct library or DLL loaded into memory first.".to_string())
-		}
-	}
-	
-	pub fn is_null(&self) -> bool {
-		self.internal_ptr.is_null()
-	}
-}
-
-#[repr(C)]
-pub struct CLRRuntimeHost {
-	internal_ptr: *mut c_void,
-	started: bool, 
-	assemblies: Vec<Assembly>
-}
-
-impl Drop for CLRRuntimeHost {
-	fn drop(&mut self) {
-		if self.started {
-			unsafe {
-				CLRRuntimeHost_stop(self.internal_ptr);
-			}
-		}
-	}
-}
-
-impl CLRRuntimeHost {
-	pub fn start(&mut self) -> bool {
-		if !self.started {
-			self.started = unsafe {
-				CLRRuntimeHost_start(self.internal_ptr)
-			};
-		}
-		self.started
-	}
-	
-	pub fn is_null(&self) -> bool {
-		self.internal_ptr.is_null()
-	}
-	
-	pub fn load_assembly(&mut self, runtime_info: CLRRuntimeInfo, assembly_name: &str) -> bool {
-		if !self.started {
-			self.start();
-		}
-		let cs_assembly_name = CString::new(assembly_name).unwrap();
-		let res = unsafe {
-			CLRRuntimeHost_load_assembly(self.internal_ptr, runtime_info.internal_ptr, cs_assembly_name.as_ptr())
-		};
-		if res.ok {
-			self.assemblies.push(Assembly::new(res.c_ptr));
-		}
-		return res.ok;
-	}
-}
 
 #[cfg(test)]
 mod tests {
